updates gguf conversion pattern; allows for hugging face repo overrides to be set in scripts
