+++ py-gguf/convert_t5_encoder_to_gguf
@@ -1,0 +1,24 @@
+#!/usr/bin/env python3
+
+import argparse
+from tts_encoders import T5Encoder, DEFAULT_PARLER_REPO_MINI_ID, DEFAULT_PARLER_REPO_LARGE_ID
+from os.path import isdir, dirname
+
+
+def parse_arguments():
+    parser = argparse.ArgumentParser()
+    parser.add_argument("--save-path", type=str, required=True, help="The path to save the converted GGUF tts model too.")
+    parser.add_argument("--large-model", default=False, action='store_true', help="Whether to encode the large version of Parler TTS.")
+    parser.add_argument("--repo-id-override", type=str, required=False, help="A custom Huggingface repository to pull the model from.")
+    parser.add_argument("--never-make-dirs", default=False, action="store_true", help="When set the script will never add new directories.")
+    return parser.parse_known_args()
+
+
+if __name__ == '__main__':
+    args, _ = parse_arguments()
+    if not isdir(dirname(args.save_path)) and args.never_make_dirs:
+        raise ValueError(f"model path, {args.save_path} is not a valid path.")
+    repo_id = DEFAULT_PARLER_REPO_LARGE_ID if args.large_model else DEFAULT_PARLER_REPO_MINI_ID
+    if args.repo_id_override:
+        repo_id = args.repo_id_override
+    T5Encoder(args.save_path, repo_id=repo_id).write()
